moved the data display functions into a new module and updated imports
