+++ plotting/plot_handler.py
@@ -25,7 +25,7 @@ from data.data_collector import prepare_bands_info, prepare_wannier_info, prepar
 from data.models import ProjectSetup
 from core.project_setup import initialize_project
 from plotting.plotters import BandPlotter, WannierComparePlotter, DOSPlotter
-from ui.plot_info_display import display_band_plot_info, display_wannier_plot_info, display_pdos_plot_info
+from ui.display_plot_info import display_band_plot_info, display_wannier_plot_info, display_pdos_plot_info
 from ui.ui_helpers import print_header, console, print_success, prompt_input
 from ui.print_thanks import print_animated_ascii
