+++ ui/display_plot_info.py
